Update three-to-ammo dependency to latest 1.0.1
This version deeply reforms the api to enable use with workers. Now creating collision requires pre-calculating vertices, indexes and matrices so that they are expressed in basic Javascript types.
The new version is also written using the "include" and "export" syntax, so an additional dependency to esmify has been introduced and the command lines and tests updated to use esmify on the codebase.
The patch previously applied to three-to-ammo by us is obsolete, as we can pass the options to the entire api directly in our codebase.